move stock packages and splitters into stock directory
