+++ core/app/models/spree/stock/packer.rb
@@ -14,7 +14,7 @@ module Spree
       end
 
       def default_package
-        package = StockPackage.new(stock_location)
+        package = Package.new(stock_location)
         order.line_items.each do |line_item|
           on_hand, backordered = stock_status(line_item.variant, line_item.quantity)
           package.add line_item.variant, on_hand, :on_hand if on_hand > 0
@@ -25,10 +25,10 @@ module Spree
 
       def splitter
         # TODO build a chain of splitters
-        StockSplitter::Base.new(stock_location, order)
+        Splitter::Base.new(stock_location, order)
       end
 
-      # private
+      private
       def stock_status(variant, quantity)
         item = stock_location.stock_item(variant)
