+++ core/app/models/spree/stock/splitter.rb
@@ -1,0 +1,25 @@
+module Spree
+  module Stock
+    module Splitter
+      class Base
+        attr_accessor :stock_location, :order
+
+        def initialize(stock_location, order)
+          @stock_location = stock_location
+          @order = order
+        end
+
+        def split(packages)
+          packages
+        end
+      end
+
+      class ShippingCategory < Base
+        def split(packages)
+          #TODO regroup items by shipping category
+          packages
+        end
+      end
+    end
+  end
+end
